Add H264 to CameraServer
The existing function was redone as a class, then H264CameraServer was made in its image. The CameraServer function is essentially a wrapper around one of the two now, chose on on_start.

H264CameraServer sacrifices the first frame as a way to determine image dimensions and is locked to 30 FPS for safety. (I have no idea what happens if you make target FPS higher than reality.)